# docs/design/pages/openapi.json
{
  "components": {
    "schemas": {
      "APIResponse_AgentConfigOut_": {
        "properties": {
          "_links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "$ref": "#/components/schemas/AgentConfigOut"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Meta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "data"
        ],
        "title": "APIResponse[AgentConfigOut]",
        "type": "object"
      },
...
      "APIResponse_ConfigOut_": {
        "properties": {
          "_links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ConfigOut"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Meta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "data"
        ],
        "title": "APIResponse[ConfigOut]",
        "type": "object"
      },
...
      "APIResponse_Dict_str__float__": {
        "properties": {
          "_links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Data",
            "type": "object"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Meta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "data"
        ],
        "title": "APIResponse[Dict[str, float]]",
        "type": "object"
      },
...
      "APIResponse_RiskConfigOut_": {
        "properties": {
          "_links": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Links"
              },
              {
                "type": "null"
              }
            ]
          },
          "data": {
            "$ref": "#/components/schemas/RiskConfigOut"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Meta"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "data"
        ],
        "title": "APIResponse[RiskConfigOut]",
        "type": "object"
      },
...
      "AgentConfigOut": {
        "description": "Full agent detail including all static configurable parameters.",
        "properties": {
          "cycles": {
            "title": "Cycles",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "implemented": {
            "title": "Implemented",
            "type": "boolean"
          },
          "last_action_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action At"
          },
          "params": {
            "title": "Params",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "domain",
          "implemented",
          "description",
          "status",
          "cycles"
        ],
        "title": "AgentConfigOut",
        "type": "object"
      },
...
      "AlertsConfigPatch": {
        "properties": {
          "euphoria_size_multiplier": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Euphoria Size Multiplier"
          },
          "overconfidence_margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overconfidence Margin"
          },
          "revenge_size_multiplier": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revenge Size Multiplier"
          },
          "risk_of_ruin_alert_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Of Ruin Alert Pct"
          }
        },
        "title": "AlertsConfigPatch",
        "type": "object"
      },
...
      "ConfigOut": {
        "properties": {
          "app_env": {
            "title": "App Env",
            "type": "string"
          },
          "autonomy_level": {
            "title": "Autonomy Level",
            "type": "integer"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "exchange": {
            "title": "Exchange",
            "type": "string"
          },
          "initial_capital": {
            "title": "Initial Capital",
            "type": "number"
          },
          "orchestrator_interval_seconds": {
            "title": "Orchestrator Interval Seconds",
            "type": "integer"
          }
        },
        "required": [
          "exchange",
          "dry_run",
          "initial_capital",
          "orchestrator_interval_seconds",
          "autonomy_level",
          "app_env"
        ],
        "title": "ConfigOut",
        "type": "object"
      },
      "ConfigPatch": {
        "properties": {
          "initial_capital": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Capital"
          },
          "orchestrator_interval_seconds": {
            "anyOf": [
              {
                "maximum": 3600.0,
                "minimum": 5.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orchestrator Interval Seconds"
          }
        },
        "title": "ConfigPatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Links": {
        "properties": {
          "related": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related"
          },
          "self": {
            "title": "Self",
            "type": "string"
          }
        },
        "required": [
          "self"
        ],
        "title": "Links",
        "type": "object"
      },
...
      "Meta": {
        "properties": {
          "page": {
            "default": 1,
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "default": 20,
            "title": "Per Page",
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "Meta",
        "type": "object"
      },
...
      "RiskConfigOut": {
        "properties": {
          "circuit_breaker_consecutive_losses": {
            "title": "Circuit Breaker Consecutive Losses",
            "type": "integer"
          },
          "circuit_breaker_enabled": {
            "title": "Circuit Breaker Enabled",
            "type": "boolean"
          },
          "circuit_breaker_trigger_pct": {
            "title": "Circuit Breaker Trigger Pct",
            "type": "number"
          },
          "cooldown_hours": {
            "title": "Cooldown Hours",
            "type": "integer"
          },
          "kelly_fraction": {
            "title": "Kelly Fraction",
            "type": "number"
          },
          "max_daily_loss_pct": {
            "title": "Max Daily Loss Pct",
            "type": "number"
          },
          "max_monthly_loss_pct": {
            "title": "Max Monthly Loss Pct",
            "type": "number"
          },
          "max_position_size_pct": {
            "title": "Max Position Size Pct",
            "type": "number"
          },
          "max_weekly_loss_pct": {
            "title": "Max Weekly Loss Pct",
            "type": "number"
          },
          "min_position_size_pct": {
            "title": "Min Position Size Pct",
            "type": "number"
          },
          "min_risk_reward_ratio": {
            "title": "Min Risk Reward Ratio",
            "type": "number"
          },
          "stop_loss_default_pct": {
            "title": "Stop Loss Default Pct",
            "type": "number"
          },
          "stop_loss_max_pct": {
            "title": "Stop Loss Max Pct",
            "type": "number"
          },
          "take_profit_default_pct": {
            "title": "Take Profit Default Pct",
            "type": "number"
          }
        },
        "required": [
          "max_position_size_pct",
          "min_position_size_pct",
          "stop_loss_default_pct",
          "stop_loss_max_pct",
          "take_profit_default_pct",
          "min_risk_reward_ratio",
          "max_daily_loss_pct",
          "max_weekly_loss_pct",
          "max_monthly_loss_pct",
          "circuit_breaker_enabled",
          "circuit_breaker_trigger_pct",
          "circuit_breaker_consecutive_losses",
          "cooldown_hours",
          "kelly_fraction"
        ],
        "title": "RiskConfigOut",
        "type": "object"
      },
      "RiskConfigPatch": {
        "properties": {
          "confirm": {
            "default": false,
            "description": "Obrigat\u00f3rio confirm=true para gravar altera\u00e7\u00f5es de risco",
            "title": "Confirm",
            "type": "boolean"
          },
          "kelly_fraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kelly Fraction"
          },
          "max_daily_loss_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Daily Loss Pct"
          },
          "max_monthly_loss_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Monthly Loss Pct"
          },
          "max_position_size_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Position Size Pct"
          },
          "max_weekly_loss_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Weekly Loss Pct"
          },
          "stop_loss_default_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss Default Pct"
          },
          "take_profit_default_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit Default Pct"
          }
        },
        "title": "RiskConfigPatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Gateway de orquestra\u00e7\u00e3o de trading com agentes AI.",
    "title": "Criptotrade API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agents/{agent_id}/config": {
      "get": {
        "operationId": "get_agent_config_v1_agents__agent_id__config_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AgentConfigOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Configura\u00e7\u00e3o completa de um agente (par\u00e2metros est\u00e1ticos; funciona para stubs)",
        "tags": [
          "agents"
        ]
      },
      "patch": {
        "operationId": "patch_agent_config_v1_agents__agent_id__config_patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Params",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_AgentConfigOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualiza par\u00e2metros de um agente",
        "tags": [
          "config"
        ]
      }
    },
...
    "/v1/alerts/config": {
      "patch": {
        "operationId": "patch_alerts_config_v1_alerts_config_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertsConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_Dict_str__float__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualiza thresholds do behavioral guard",
        "tags": [
          "config"
        ]
      }
    },
...
    "/v1/config": {
      "get": {
        "operationId": "get_config_v1_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ConfigOut_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Configura\u00e7\u00f5es gerais do sistema",
        "tags": [
          "config"
        ]
      },
      "patch": {
        "operationId": "patch_config_v1_config_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ConfigOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualiza configura\u00e7\u00f5es gerais (in-memory para MVP)",
        "tags": [
          "config"
        ]
      }
    },
...
    "/v1/risk/config": {
      "get": {
        "operationId": "get_risk_config_v1_risk_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_RiskConfigOut_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Par\u00e2metros de risco (l\u00ea risk_params.yaml)",
        "tags": [
          "risk"
        ]
      },
      "patch": {
        "operationId": "patch_risk_config_v1_risk_config_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_RiskConfigOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualiza par\u00e2metros de risco (grava em risk_params.yaml)",
        "tags": [
          "risk"
        ]
      }
    },
...
  }
}